+++ provision/clamav.sh
@@ -332,6 +332,7 @@ test_clamav()
 	echo "It works! (clamd is listening)"
 }
 
+tell_settings CLAMAV
 base_snapshot_exists || exit 1
 migrate_clamav_dbs
 create_staged_fs clamav